+++ docs/release_notes.rst
@@ -3,6 +3,29 @@
 Release notes
 =============
 
+Version 0.3
+-----------
+
+This version is a minor release with critical but non trivial fixes before
+version 1.0.0
+
+Improvements
+++++++++++++
+
+- Don't remove timestamp from final products when no concatenation is done
+  (`#69 <https://gitlab.orfeo-toolbox.org/s1-tiling/s1tiling/-/issues/69>`_).
+  Up to now timetag was always changed to ``txxxxxx``
+
+Bug fixed
++++++++++
+
+Optimizations
++++++++++++++
+
+- Downloading and unzipping of Sentinel-1 products is done in parallel
+  (`!31 <https://gitlab.orfeo-toolbox.org/s1-tiling/s1tiling/-/merge_requests/31>`_)
+
+
 Version 0.2
 -----------
 
@@ -14,7 +37,7 @@ Improvements
 
 - Provide the possibility to use linear interpolation for orthorectification step
 - Support OTB 7.3
-- Various return code after execution are now provided ((`#72 <https://gitlab.orfeo-toolbox.org/s1-tiling/s1tiling/-/issues/72>`_)
+- Various return code after execution are now provided (`#72 <https://gitlab.orfeo-toolbox.org/s1-tiling/s1tiling/-/issues/72>`_)
 - Improved architecture to help maintenance
 - Project stability has been improved